+++ dashboard/views/custom-products/form.php
@@ -178,7 +178,7 @@ if (0 < $recordId) {
                                             <strong> <?php echo Labels::getLabel('LBL_UPLOAD_IMAGES(S)', $langId); ?></strong>
                                             <span class="text-muted form-text"><?php echo Labels::getLabel('MSG_PNG,JPEG_ACCEPTED', $langId); ?></span></button></li>
                                 </ul>
-                                <div class="form-text text-muted pt-2"><?php echo Labels::getLabel('MSG_PAY_ATTENTION_TO_THE_PICTURE_QUALITY', $langId); ?></div>
+                                <div class="form-text text-muted pt-2"><?php echo sprintf(Labels::getLabel('LBL_Preferred_Dimensions_%s', $siteLangId), $imgFrm->getField('min_width')->value . ' x ' . $imgFrm->getField('min_height')->value); ?></div>
                             </div>
                         </div>
                     </div>
@@ -286,7 +286,7 @@ if (0 < $recordId) {
                             <div class="card-foot">
                                 <div class="row justify-content-end">
                                     <div class="col-auto">
-                                        <button class="btn btn-icon btn-outline-gray btn-sm" onclick="digitalDownloadsForm(<?php echo applicationConstants::DIGITAL_DOWNLOAD_FILE; ?>);">
+                                        <button class="btn btn-icon btn-outline-gray btn-sm" type="button" onclick="digitalDownloadsForm(<?php echo applicationConstants::DIGITAL_DOWNLOAD_FILE; ?>);">
                                             <svg class="svg btn-icon-start" width="18" height="18">
                                                 <use xlink:href="<?php echo CONF_WEBROOT_URL; ?>images/retina/sprite-actions.svg#add">
                                                 </use>
@@ -316,7 +316,7 @@ if (0 < $recordId) {
                             <div class="card-foot">
                                 <div class="row justify-content-end">
                                     <div class="col-auto">
-                                        <button class="btn btn-icon btn-outline-gray btn-sm" onclick="digitalDownloadsForm(<?php echo applicationConstants::DIGITAL_DOWNLOAD_LINK; ?>);">
+                                        <button class="btn btn-icon btn-outline-gray btn-sm" type="button" onclick="digitalDownloadsForm(<?php echo applicationConstants::DIGITAL_DOWNLOAD_LINK; ?>);">
                                             <svg class="svg btn-icon-start" width="18" height="18">
                                                 <use xlink:href="<?php echo CONF_WEBROOT_URL; ?>images/retina/sprite-actions.svg#add">
                                                 </use>