+++ dashboard/views/products/get-digital-download-links.php
@@ -3,19 +3,19 @@
 if (!$showNoRecordFound && empty($links)) {
     return;
 }
-$arr_flds = [];
-
+$arr_flds = [
+ 'listSerial' =>  Labels::getLabel('LBL_#', $siteLangId),
+ 'pdl_download_link' =>  Labels::getLabel('LBL_DOWNLOAD_LINK', $siteLangId),
+ 'pdl_preview_link' =>  Labels::getLabel('LBL_PREVIEW_LINK', $siteLangId),
+ 'action' =>  Labels::getLabel('LBL_ACTION_BUTTONS', $siteLangId),
+];
 if ($canDo) {
-    $arr_flds['pdl_download_link'] = Labels::getLabel('LBL_DOWNLOAD_LINK', $siteLangId);
-}
-$arr_flds['pdl_preview_link'] = Labels::getLabel('LBL_PREVIEW_LINK', $siteLangId);
-if ($canDo) {
-    $arr_flds['action'] = Labels::getLabel('LBL_ACTION_BUTTONS', $siteLangId);
+    unset($arr_flds['listSerial']);
+}else{
+    unset($arr_flds['action'],$arr_flds['pdl_download_link']);  
 }
 
-
-
-$tbl = new HtmlElement('table', array('width' => '100%', 'class' => 'table table-justified'));
+$tbl = new HtmlElement('table', array('width' => '100%', 'class' => 'table '. (isset($arr_flds['action']) ? 'table-justified': '')));
 $th = $tbl->appendElement('thead')->appendElement('tr', array('class' => 'hide--mobile'));
 foreach ($arr_flds as $key => $val) {
     $tdAttr = ('action' == $key) ? ['class' => 'align-right', 'width' => '20%'] : ['width' => '40%'];